+++ CHANGES.md
@@ -11,6 +11,10 @@ Use this section to keep track of changes in the works.
 ### Changed
 
 ### Fixed
+* Fix orbit generation to play niceley with changes in pymatgen
+  `Structure.sites_in_sphere` return value.
+  [\#125](https://github.com/CederGroupHub/smol/pull/125)
+  ([lbluque](https://github.com/lbluque))
 
 ## [v1.0.1](https://github.com/CederGroupHub/smol/tree/v1.0.1) (2021-03-03)
 #### [Full Changelog](https://github.com/CederGroupHub/smol/compare/v1.0.0...v1.0.1)